+++ darkside-tests/Cargo.toml
@@ -9,7 +9,7 @@ chain_generic_tests = []
 
 [dependencies]
 zingolib = { path = "../zingolib", features = ["darkside_tests"] }
-zingo-sync = { path = "../zingo-sync" }
+pepper-sync = { path = "../pepper-sync" }
 zingo-netutils = { path = "../zingo-netutils" }
 zingo-status = { path = "../zingo-status" }
 testvectors = { path = "../testvectors" }
