+++ pepper-sync/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "zingo-sync"
+name = "pepper-sync"
 version = "0.1.0"
 edition = "2021"
 
 [features]
 default = [ "wallet_essentials" ]
-# utility not necessary for the sync engine but essential for wallets using zingo_sync::primitives for wallet functionality
+# utility not necessary for the sync engine but essential for wallets using pepper_sync::primitives for wallet functionality
 wallet_essentials = []
 
 [dependencies]
