+++ pepper-sync/src/error.rs
