+++ pepper-sync/src/scan/compact_blocks.rs
