+++ pepper-sync/src/scan/error.rs
