+++ pepper-sync/src/wallet/traits.rs
