+++ zingolib/Cargo.toml
@@ -18,7 +18,7 @@ zaino-test = ['test-elevation']
 zingo-memo = { path = "../zingo-memo" }
 zingo-status = { path = "../zingo-status" }
 zingo-netutils = { path = "../zingo-netutils" }
-zingo-sync = { path = "../zingo-sync" }
+pepper-sync = { path = "../pepper-sync" }
 testvectors = { path = "../testvectors", optional = true }
 
 orchard = { workspace = true }
