+++ zingolib/src/testutils/assertions.rs
@@ -2,9 +2,9 @@
 
 use nonempty::NonEmpty;
 
+use pepper_sync::wallet::WalletTransaction;
 use zcash_client_backend::proposal::{Proposal, Step};
 use zcash_primitives::transaction::TxId;
-use zingo_sync::wallet::WalletTransaction;
 
 use crate::{lightclient::LightClient, wallet::LightWallet};
