+++ zingolib/src/wallet/disk.rs
@@ -294,10 +294,10 @@ impl LightWallet {
             price: Arc::new(RwLock::new(price)),
             wallet_blocks: BTreeMap::new(),
             wallet_transactions: HashMap::new(),
-            nullifier_map: zingo_sync::wallet::NullifierMap::new(),
+            nullifier_map: pepper_sync::wallet::NullifierMap::new(),
             outpoint_map: BTreeMap::new(),
-            shard_trees: zingo_sync::wallet::ShardTrees::new(),
-            sync_state: zingo_sync::wallet::SyncState::new(),
+            shard_trees: pepper_sync::wallet::ShardTrees::new(),
+            sync_state: pepper_sync::wallet::SyncState::new(),
             transparent_addresses: BTreeMap::new(),
             unified_addresses: AppendOnlyVec::new(),
             network,
