+++ zingolib/src/wallet/keys.rs
@@ -3,6 +3,7 @@
 //! from a source outside of the code-base e.g. a wallet-file.
 use crate::config::ZingoConfig;
 use base58::ToBase58;
+use pepper_sync::keys::transparent::{self, TransparentAddressId, TransparentScope};
 use sapling_crypto::{
     zip32::{DiversifiableFullViewingKey, ExtendedSpendingKey},
     PaymentAddress,
@@ -13,7 +14,6 @@ use zcash_keys::address::UnifiedAddress;
 use zcash_primitives::{
     consensus::NetworkConstants, legacy::TransparentAddress, zip32::ChildIndex,
 };
-use zingo_sync::keys::transparent::{self, TransparentAddressId, TransparentScope};
 
 use super::{error::KeyError, LightWallet};
