+++ zingolib/src/wallet/sync.rs
@@ -2,15 +2,15 @@
 
 use std::collections::{BTreeMap, HashMap};
 
-use zcash_keys::keys::UnifiedFullViewingKey;
-use zcash_primitives::consensus::BlockHeight;
-use zingo_sync::{
+use pepper_sync::{
     keys::transparent::TransparentAddressId,
     wallet::traits::{
         SyncBlocks, SyncNullifiers, SyncOutPoints, SyncShardTrees, SyncTransactions, SyncWallet,
     },
     wallet::{Locator, NullifierMap, OutputId, ShardTrees, SyncState, WalletBlock},
 };
+use zcash_keys::keys::UnifiedFullViewingKey;
+use zcash_primitives::consensus::BlockHeight;
 use zip32::AccountId;
 
 use super::{error::WalletError, LightWallet};
@@ -73,7 +73,7 @@ impl SyncTransactions for LightWallet {
     fn get_wallet_transactions(
         &self,
     ) -> Result<
-        &HashMap<zcash_primitives::transaction::TxId, zingo_sync::wallet::WalletTransaction>,
+        &HashMap<zcash_primitives::transaction::TxId, pepper_sync::wallet::WalletTransaction>,
         Self::Error,
     > {
         Ok(&self.wallet_transactions)
@@ -82,7 +82,7 @@ impl SyncTransactions for LightWallet {
     fn get_wallet_transactions_mut(
         &mut self,
     ) -> Result<
-        &mut HashMap<zcash_primitives::transaction::TxId, zingo_sync::wallet::WalletTransaction>,
+        &mut HashMap<zcash_primitives::transaction::TxId, pepper_sync::wallet::WalletTransaction>,
         Self::Error,
     > {
         Ok(&mut self.wallet_transactions)
