+++ zingolib/src/wallet/transaction.rs
@@ -1,5 +1,5 @@
+use pepper_sync::wallet::{OutputId, OutputInterface, TransparentCoin, WalletTransaction};
 use zcash_primitives::transaction::components::Amount;
-use zingo_sync::wallet::{OutputId, OutputInterface, TransparentCoin, WalletTransaction};
 
 use super::{
     error::{FeeError, KindError},
